requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
bandit==1.9.4
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
ctranslate2==4.7.1
faster-whisper==1.2.1
filelock==3.25.0
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.4.2
ollama==0.6.1
onnxruntime==1.24.2
packaging==26.0
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyYAML==6.0.3
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
sounddevice==0.5.5
stevedore==5.7.0
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
webrtcvad-wheels==2.0.14
